% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexicmstate.R
\name{flexicmstate}
\alias{flexicmstate}
\title{flexicmstate}
\description{
Flexible modelling of interval-censored Multi-state models
}
\details{
_PACKAGE package
}
\author{
Daniel Gomon <dgstatsoft at gmail.com>
}
